pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mini-etl"
version = "0.1.0"
description = "A lightweight, streaming ETL system in Python"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Adjei" }
]
keywords = ["etl", "data-pipeline", "streaming", "pandas", "data-engineering"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
]

dependencies = [
    "pandas>=2.0.0",
    "sqlalchemy>=2.0.0",
    "pyarrow>=14.0.0",
    "click>=8.0.0",
    "tqdm>=4.60.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "openpyxl>=3.1.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
ui = [
    "streamlit>=1.28.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "httpx>=0.25.0",
    "responses>=0.24.0",
]
all = [
    "mini-etl[ui,dev]",
]

[project.scripts]
mini-etl = "mini_etl.cli:main"

[project.urls]
Homepage = "https://github.com/adjei/mini-etl"
Repository = "https://github.com/adjei/mini-etl"
Issues = "https://github.com/adjei/mini-etl/issues"

[tool.hatch.build.targets.wheel]
packages = ["mini_etl"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.coverage.run]
source = ["mini_etl"]
omit = ["mini_etl/ui/*", "tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
